update graduateProject API
