+++ mu-plugins/search-route.php
@@ -19,76 +19,7 @@
       return $results;
    }
 
-   function graduateProjectSearchResults($data) {
-      
-      $mainQuery = new WP_Query(array(
-         'post_type' => 'graduateProjects',
-         's' => sanitize_text_field($data['term'])
-      ));
-      
-      $gradRange = [100, 150];
-
-      $results = array();
-      for($i = $gradRange[1]; $i >= $gradRange[0]; $i--){
-         array_push($results,array(
-            'sortTitle' => $i,
-            'sortList' => array()
-         ));
-      }
-
-      while($mainQuery->have_posts()) {
-         $mainQuery->the_post();
-
-         $sortTitle = get_field('sortTitle');
-
-         $collection = array(
-            'id' => get_the_ID(),
-            'workTitle' => get_field('workTitle'),
-            'workImgUrl' => get_field('workImgUrl')['url'],
-            'introduction' => get_field('introduction'),
-            'author' => get_field('author'),
-            'instructor' => get_field('instructor'),
-            'honor' => get_field('honor'),
-            'relatedLinks' => array(
-               array(
-                  'linkTitle' => get_field('relatedLinks')['link1']['linkTitle'],
-                  'linkUrl' => get_field('relatedLinks')['link1']['linkUrl'],
-               ),  
-               array(
-                  'linkTitle' => get_field('relatedLinks')['link2']['linkTitle'],
-                  'linkUrl' => get_field('relatedLinks')['link2']['linkUrl'],
-               ),  
-               array(
-                  'linkTitle' => get_field('relatedLinks')['link3']['linkTitle'],
-                  'linkUrl' => get_field('relatedLinks')['link3']['linkUrl'],
-               ),  
-               array(
-                  'linkTitle' => get_field('relatedLinks')['link4']['linkTitle'],
-                  'linkUrl' => get_field('relatedLinks')['link4']['linkUrl'],
-               ),  
-            )
-         );
-
-         for($i = count($collection['relatedLinks']) - 1; $i >= 0; $i--){
-            if($collection['relatedLinks'][$i]['linkUrl'] == ''){
-               unset($collection['relatedLinks'][$i]);
-            }
-         }
-
-         array_push($results[$gradRange[1] - $sortTitle]['sortList'], $collection);
-      }
-
-      for($i = $gradRange[1] - $gradRange[0]; $i >= 0; $i--){
-         if(count($results[$i]['sortList']) == 0){
-            unset($results[$i]);
-            $results = array_values($results);
-         }
-      }
-      
-      return $results;
-   }
-
-
+  
    function staffSearchResults($data) {
       $mainQuery = new WP_Query(array(
          'post_type' => 'staff',
@@ -131,11 +62,11 @@
    
             $groupID = 0;
             switch(get_field('grouptitle')){
-               case "專任老師/互動科技領域":
+               case "互動科技領域":
                   $groupID = 0; break;
-               case "專任老師/遊戲設計領域":
+               case "遊戲設計領域":
                   $groupID = 1; break;
-               case "專任老師/創意設計領域":
+               case "創意設計領域":
                   $groupID = 2; break;
                case "兼任老師":
                   $groupID = 3; break;
